+++ publisher-domain/src/main/java/com/example/publisher/application/business/exception/PublisherNotFoundException.java
@@ -1,14 +1,14 @@
 package com.example.publisher.application.business.exception;
 
 public class PublisherNotFoundException extends RuntimeException {
-    private final int publisherId;
+    private final long publisherId;
 
-    public PublisherNotFoundException(String message, int publisherId) {
+    public PublisherNotFoundException(String message, long publisherId) {
         super(message);
         this.publisherId = publisherId;
     }
 
-    public int getPublisherId() {
+    public long getPublisherId() {
         return publisherId;
     }
 }
